requirements.txt
psycopg[binary]==3.3.2
click==8.1.7
python-dotenv==1.0.1
pytest==8.0.0
